Move common files, adjust namespaces
