+++ ChartJs.Blazor/ChartJS/Common/Axes/Axis.cs
@@ -1,6 +1,6 @@
 ﻿using ChartJs.Blazor.ChartJS.Common.Enums;
 
-namespace ChartJs.Blazor.ChartJS.Common
+namespace ChartJs.Blazor.ChartJS.Common.Axes
 {
     /// <summary>
     /// As per documentation here https://www.chartjs.org/docs/latest/axes/#common-configuration.
