+++ ChartJs.Blazor/ChartJS/Common/Axes/PointLabels.cs
@@ -2,7 +2,7 @@
 using ChartJs.Blazor.ChartJS.Common.Enums;
 using ChartJs.Blazor.Util;
 
-namespace ChartJs.Blazor.ChartJS.PolarAreaChart.Axis
+namespace ChartJs.Blazor.ChartJS.Common.Axes
 {
     /// <summary>
     /// The point labels sub-config of the linear-radial-axis-configuration (see <see cref="LinearRadialAxis"/>).
