+++ ChartJs.Blazor/ChartJS/Common/Axes/Ticks/CartesianTicks.cs
@@ -1,4 +1,4 @@
-﻿namespace ChartJs.Blazor.ChartJS.LineChart.Axes.Ticks
+﻿namespace ChartJs.Blazor.ChartJS.Common.Axes.Ticks
 {
     /// <summary>
     /// As per documentation here https://www.chartjs.org/docs/latest/axes/cartesian/#tick-configuration
