+++ ChartJs.Blazor/ChartJS/Common/Axes/Ticks/LinearCartesianTicks.cs
@@ -1,5 +1,5 @@
 ﻿
-namespace ChartJs.Blazor.ChartJS.LineChart.Axes.Ticks
+namespace ChartJs.Blazor.ChartJS.Common.Axes.Ticks
 {
     /// <summary>
     /// https://www.chartjs.org/docs/latest/axes/cartesian/linear.html#tick-configuration-options
