+++ WebCore/Pages/LineLogarithmicExample.razor
@@ -6,8 +6,8 @@
 @using ChartJs.Blazor.ChartJS.Common.Enums
 @using ChartJs.Blazor.ChartJS.Common.Legends
 @using ChartJs.Blazor.ChartJS.LineChart
-@using ChartJs.Blazor.ChartJS.LineChart.Axes
-@using ChartJs.Blazor.ChartJS.LineChart.Axes.Ticks
+@using ChartJs.Blazor.ChartJS.Common.Axes
+@using ChartJs.Blazor.ChartJS.Common.Axes.Ticks
 @using ChartJs.Blazor.Util
 @inject WeatherForecastService ForecastService
