+++ WebCore/Pages/SimpleLineLinearExample.razor
@@ -5,7 +5,7 @@
 @using ChartJs.Blazor.ChartJS.Common.Enums
 @using ChartJs.Blazor.ChartJS.Common.Legends
 @using ChartJs.Blazor.ChartJS.LineChart
-@using ChartJs.Blazor.ChartJS.LineChart.Axes
+@using ChartJs.Blazor.ChartJS.Common.Axes
 @using ChartJs.Blazor.Util
 
 <h2>Line Linear Chart</h2>
